+++ python/ray/data/tests/conftest.py
@@ -13,11 +13,11 @@ import ray
 from ray._common.test_utils import wait_for_condition
 from ray._private.internal_api import get_memory_info_reply, get_state_from_address
 from ray.data._internal.tensor_extensions.arrow import ArrowTensorArray
-from ray.data.arrow_utils import get_pyarrow_version
 from ray.data.block import BlockExecStats, BlockMetadata
 from ray.data.constants import TENSOR_COLUMN_NAME
 from ray.data.context import DEFAULT_TARGET_MAX_BLOCK_SIZE, DataContext, ShuffleStrategy
 from ray.data.tests.mock_server import *  # noqa
+from ray.data.util.arrow_utils import get_pyarrow_version
 
 # Trigger pytest hook to automatically zip test cluster logs to archive dir on failure
 from ray.tests.conftest import *  # noqa
