+++ python/ray/data/tests/expressions/test_arithmetic.py
@@ -11,9 +11,9 @@ from packaging.version import parse as parse_version
 
 import ray
 from ray.data._internal.util import rows_same
-from ray.data.arrow_utils import get_pyarrow_version
 from ray.data.expressions import col, lit
 from ray.data.tests.conftest import *  # noqa
+from ray.data.util.arrow_utils import get_pyarrow_version
 from ray.tests.conftest import *  # noqa
 
 pytestmark = pytest.mark.skipif(
