+++ python/ray/data/tests/test_groupby_e2e.py
@@ -30,12 +30,12 @@ from ray.data.aggregate import (
     Sum,
     Unique,
 )
-from ray.data.arrow_utils import get_pyarrow_version
 from ray.data.block import BlockAccessor
 from ray.data.context import DataContext, ShuffleStrategy
 from ray.data.expressions import col
 from ray.data.tests.conftest import *  # noqa
 from ray.data.tests.util import named_values
+from ray.data.util.arrow_utils import get_pyarrow_version
 from ray.tests.conftest import *  # noqa
 
 RANDOM_SEED = 123
